Move common page logic to web
