+++ web/go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/decred/dcrd/chaincfg/v2 v2.3.0
 	github.com/decred/dcrd/dcrutil/v2 v2.0.1
 	github.com/decred/dcrd/rpc/jsonrpc/types/v2 v2.3.0
-	github.com/decred/dcrdata/explorer/types v1.1.0
+	github.com/decred/dcrd/txscript/v2 v2.1.0
 	github.com/decred/slog v1.1.0
 	github.com/dustin/go-humanize v1.0.0
 	github.com/go-chi/chi v4.1.2+incompatible
